requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
environs==9.5.0
exceptiongroup==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
icalendar==5.0.7
idna==3.4
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
mysql-connector-python==8.2.0
outcome==1.2.0
packaging==23.1
playwright==1.35.0
playwright-stealth==1.0.5
protobuf==4.21.12
pydantic==1.10.9
pyee==9.0.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
schedule==1.2.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
urllib3==2.0.3
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
